+++ Src/DiKErnel.Integration/Data/TimeDependentInput.cs
@@ -26,7 +26,7 @@ namespace DiKErnel.Integration.Data
 {
     internal class TimeDependentInput : ITimeDependentInput
     {
-        public TimeDependentInput(int beginTime, int endTime, double waterLevel, double waveHeightHm0,
+        public TimeDependentInput(double beginTime, double endTime, double waterLevel, double waveHeightHm0,
                                   double wavePeriodTm10, double waveDirection)
         {
             BeginTime = beginTime;
@@ -37,9 +37,9 @@ namespace DiKErnel.Integration.Data
             WaveDirection = waveDirection;
         }
 
-        public int BeginTime { get; }
+        public double BeginTime { get; }
 
-        public int EndTime { get; }
+        public double EndTime { get; }
 
         public double WaterLevel { get; }
