+++ Src/DiKErnel.KernelWrapper.Json.Input/Data/Generic/JsonInputData.cs
@@ -25,7 +25,7 @@ namespace DiKErnel.KernelWrapper.Json.Input.Data.Generic
     internal class JsonInputData
     {
         [JsonProperty(JsonInputDefinitions.Time)]
-        public IReadOnlyList<int> Times { get; private set; }
+        public IReadOnlyList<double> Times { get; private set; }
 
         [JsonProperty(JsonInputDefinitions.HydraulicLoads)]
         public JsonInputHydraulicData HydraulicData { get; private set; }
